+++ jdbc/src/Transaction.kt
@@ -14,41 +14,40 @@ import kotlin.coroutines.CoroutineContext
 /** Disable transaction for a route or a job. This will most likely leave the connection in auto-commit mode (depending on connection pool settings). */
 @Target(FUNCTION, CLASS) annotation class NoTransaction
 
-class Transaction(val db: DataSource): AutoCloseable {
+class Transaction: AutoCloseable {
   companion object {
     private val log = logger()
     private val threadLocal = ThreadLocal<Transaction>()
     fun current(): Transaction? = threadLocal.get()
   }
 
-  private var conn: Connection? = null
-  val connection: Connection get() = conn ?: openConnection()
-
-  private fun openConnection() = db.connection.apply {
-    autoCommit = false
-    conn = this
+  private val connections = HashMap<DataSource, Connection>()
+  fun connection(db: DataSource): Connection = connections.getOrPut(db) {
+    db.connection.apply { autoCommit = false }
   }
 
   override fun close() = close(true)
   fun close(commit: Boolean = true) {
-    try {
-      conn?.apply {
-        if (!autoCommit) {
-          if (commit) commit() else rollback()
-          autoCommit = true
+    connections.forEach { (_, conn) ->
+      try {
+        conn.apply {
+          if (!autoCommit) {
+            if (commit) commit() else rollback()
+            autoCommit = true
+          }
         }
+      } catch (e: SQLException) {
+        log.error("Failed to ${if (commit) "commit" else "rollback"}", e)
+      } finally {
+        try { conn.close() } catch (e: Exception) { log.error("Failed to close $conn: $e") }
       }
-    } catch (e: SQLException) {
-      log.error("Failed to ${if (commit) "commit" else "rollback"}", e)
-    } finally {
-      try { conn?.close() } catch (e: Exception) { log.error("Failed to close $conn: $e") }
-      conn = null
-      detachFromThread()
     }
+    connections.clear()
+    detachFromThread()
   }
 
-  fun commit() = conn?.commit()
-  fun rollback() = conn?.rollback()
+  fun commit() = connections.forEach { (_, conn) -> conn.commit() }
+  fun rollback() = connections.forEach { (_, conn) -> conn.rollback() }
 
   fun attachToThread() = this.also { threadLocal.set(it) }
   fun detachFromThread() = threadLocal.remove()
@@ -69,14 +68,8 @@ interface KliteTransactionAwareDataSource : DataSource {
 
 fun <R> DataSource.withConnection(block: Connection.() -> R): R {
   val tx = Transaction.current()
-  if (tx?.db == this) return tx.connection.block()
-
-  // Check if DataSource integrates with Klite's transaction management (e.g., multitenant)
-  if (this is KliteTransactionAwareDataSource) {
-    getTransactionConnection()?.let { return it.block() }
-  }
-
-  return connection.use(block)
+  return if (tx != null) tx.connection(this).block()
+         else connection.use(block)
 }
 
 class TransactionContext(val tx: Transaction? = Transaction.current()): ThreadContextElement<Transaction?>, AbstractCoroutineContextElement(Key) {