+++ jdbc/test/TransactionContextTest.kt
@@ -3,13 +3,12 @@ package klite.jdbc
 import ch.tutteli.atrium.api.fluent.en_GB.toBeTheInstance
 import ch.tutteli.atrium.api.fluent.en_GB.toEqual
 import ch.tutteli.atrium.api.verbs.expect
-import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
 import org.junit.jupiter.api.Test
 
 class TransactionContextTest {
-  val context = TransactionContext(Transaction(mockk()))
+  val context = TransactionContext(Transaction())
 
   @Test fun `transaction is attached and detached from thread`() {
     runBlocking {
@@ -26,7 +25,7 @@ class TransactionContextTest {
       withContext(context) {
         expect(Transaction.current()).toBeTheInstance(context.tx)
 
-        val tx2 = Transaction(mockk())
+        val tx2 = Transaction()
         withContext(TransactionContext(tx2)) {
           expect(Transaction.current()).toBeTheInstance(tx2)
         }