+++ openapi/src/OpenAPI.kt
@@ -1,8 +1,11 @@
 package klite.openapi
 
 import io.swagger.v3.oas.annotations.Hidden
-import klite.*
+import klite.HttpExchange
+import klite.MimeTypes
 import klite.RequestMethod.GET
+import klite.Route
+import klite.Router
 import klite.StatusCode.Companion.OK
 import klite.html.escapeJs
 import org.intellij.lang.annotations.Language
@@ -22,7 +25,6 @@ import org.intellij.lang.annotations.Language
  */
 // Generate entities separately, and reference them. To enable TS generation
 // TODO: support @Schema(description on data classes and fields)
-context(Server)
 fun Router.openApi(path: String = "/openapi", annotations: List<Annotation> = emptyList(), swaggerUIConfig: Map<String, Comparable<*>> = emptyMap()) {
   val hidden = annotations + Hidden()
   add(Route(GET, pathParamRegexer.from("$path.json"), hidden) { generateOpenAPI() })