+++ server/src/klite/UtilDecorators.kt
@@ -1,8 +1,11 @@
 package klite
 
+import klite.StatusCode.Companion.TooManyRequests
 import java.time.Instant
 import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
+import java.util.concurrent.atomic.AtomicLong
+import kotlin.math.ceil
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.days
 
@@ -23,6 +26,7 @@ fun RouterConfig.basicAuth(realm: String = "Auth", userProvider: (name: String, 
     val (name, password) = String(auth.substringAfter(" ").base64Decode()).split(':', limit = 2)
     userProvider(name, Password(password))?.let {
       e.attr("user", it)
+      e.attrPut(it)
       return@before
     }
   }
@@ -51,3 +55,31 @@ fun HttpExchange.checkLastModified(at: Instant) {
 fun HttpExchange.lastModified(at: Instant): String = DateTimeFormatter.RFC_1123_DATE_TIME.format(at.atOffset(ZoneOffset.UTC)).also {
   header("Last-Modified", it)
 }
+
+fun RouterConfig.rateLimit(limit: Int, window: Duration) {
+  val limits = Cache<String, RateLimit>(expiration = window * 3, prolongOnAccess = true)
+  val rate = limit.toDouble() / window.inWholeNanoseconds
+  val maxTokens = limit.toDouble()
+  val requestsLimited = AtomicLong()
+  Metrics.register("requestsLimited") { requestsLimited.get() }
+
+  decorator { e, handler ->
+    val limiter = limits.getOrSet(e.remoteAddress) { RateLimit(maxTokens, System.nanoTime()) }
+    synchronized(limiter) {
+      val now = System.nanoTime()
+      val refill = (now - limiter.lastRefill) * rate
+      limiter.tokens = (limiter.tokens + refill).coerceAtMost(maxTokens)
+      limiter.lastRefill = now
+      if (limiter.tokens >= 1) limiter.tokens -= 1
+      else {
+        requestsLimited.incrementAndGet()
+        val retryAfter = ceil((1 - limiter.tokens) / rate / 1e9).toInt()
+        e.header("Retry-After", retryAfter.toString())
+        throw StatusCodeException(TooManyRequests)
+      }
+    }
+    handler(e)
+  }
+}
+
+private data class RateLimit(var tokens: Double, var lastRefill: Long)